Update Firebase config, so prod and dev share same project
Generally speaking, it's not advisable to have dev and prod builds share
a same Firebase project.

But, it can be argued that, in this case, the risk of sharing it for
this project is fairly small. The reasoning is that the app uses
Firebase for analytics only, even though it makes use of push
notifications, the Firebase Cloud Messaging (FCM) only serves as a
courier, the actual management and dispatching of the notifications are
performed by OneSignal which *differentiates* DEV and PROD.