Do not clear cookies after questionnaire
The original intent of clearing cookies was to enable users to register
a different account, when they've already gone through the questionnaire.

However, that cookie clearing prevents the app from redirecting users
back to the questionnaire page to nudge them to subscribe, as they would
have to start over due to the webview forgetting all the progress.

Given the vast majority of users will only use one account and will not
want to switch accounts in that situation, we opted to just not clear
the cookies after the questionnaire, and only do so when the user
explicitly logs out.

REF: https://moacreative.teamwork.com/app/tasks/38414090