Fix issue with posting activity with media attachment
Previously, the media attachment was retrieved from the upload controller
at the time of posting, but the media at the time may be empty for some
reason.

This commit fixes the issue by retrieving the media after the upload
controller has finished uploading, and stores the media in the state
to be used when posting the activity.